feat(document-readers): move file positions
