+++ community/document-readers/email-document-reader/src/main/java/com/alibaba/cloud/ai/reader/email/eml/EmailElement.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.cloud.ai.reader.email;
+package com.alibaba.cloud.ai.reader.email.eml;
 
 import java.time.ZonedDateTime;
 import java.util.HashMap;
